fixed sql queries

+++ src/main/java/org/magenta/test/task/service/CalculateService.java
@@ -2,5 +2,5 @@ package org.magenta.test.task.service;
 
 public interface CalculateService {
     Double crowflightFormula(String from, String to);
-    Double distanceMatrix(String from, String to);
+    Integer distanceMatrix(String from, String to);
 }
